# Refactor graph sparse optimizer api