Address https://github.com/Oneflow-Inc/OneTeam/issues/1101