@ShawnXuan 2022-03-03
测试结果：
使用该分支对wdl网络在onebrain上进行了回归测试，关注的是latency，结果如下：
gpu | batch_size | baseline(ms) | #7658(ms)
-- | -- | -- | --
n1g1 | 512 | 1.132 | 1.079
n1g1 | 1024 | 1.507 | 1.488
n1g1 | 2048 | 2.425 | 2.397
n1g1 | 4096 | 4.259 | 4.228
n1g1 | 8192 | 7.634 | 7.605
n1g1 | 16384 | 14.381 | 14.301
n1g1 | 32768 | 27.141 | 27.064
n1g1 | 65536 | 54.245 | 54.328
n1g1 | 131072 | 108.502 | 108.416
n1g1 | 262144 | 226.267 | 225.76
n1g1 | 524288 | 452.222 | 447.761
n1g8 | 512 | 3.018 | 2.674
n1g8 | 1024 | 2.938 | 2.764
n1g8 | 2048 | 3.023 | 2.733
n1g8 | 4096 | 3.254 | 3.009
n1g8 | 8192 | 3.96 | 3.416
n1g8 | 16384 | 5.143 | 4.547
n1g8 | 32768 | 7.831 | 6.97
n1g8 | 65536 | 13.391 | 11.518
n1g8 | 131072 | 25.56 | 20.996
n1g8 | 262144 | 49.156 | 42.287
n1g8 | 524288 | 95.505 | 94.024


模型脚本 modles：[PR#314](https://github.com/Oneflow-Inc/models/pull/314) 
测试脚本 DLPerf: https://github.com/Oneflow-Inc/DLPerf/blob/master/OneFlow/ClickThroughRate/WideDeepLearning/scripts/bsz_x2.sh